tasks/load-odr/requirements.txt
weaviate-client
pyyaml
jmespath
boto3